Changed the symbol command to use uintptr_t so it SHOULD work in either 64-bit or 32-bit platforms
git-svn-id: https://trac.beirdo.ca:4443/svn/beirdobot/trunk@334 c66dcb5a-e80b-0410-ab0e-bb6a4f3ddef9

+++ bot/src/plugins/core.c
@@ -36,6 +36,7 @@
 #include <time.h>
 #include <mysql.h>
 #include <execinfo.h>
+#include <stdint.h>
 #include "botnet.h"
 #include "structs.h"
 #include "protos.h"
@@ -299,7 +300,7 @@ char *botHelpNotice( void *tag )
 void botCmdSymbol( IRCServer_t *server, IRCChannel_t *channel, char *who, 
                    char *msg, void *tag )
 {
-    long long int   addr;
+    uintptr_t       addr;
     void           *array[1];
     char          **strings;
